backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
python-dotenv
passlib[bcrypt]
bcrypt==3.2.0
python-jose[cryptography]
email-validator
python-multipart
alembic
psycopg2-binary